requirements.txt
Django>=4.2
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
python-docx>=1.1
natasha>=1.6
sentence-transformers>=2.2
markdown>=3.5
pytest>=7.4
pytest-django>=4.5
ruff>=0.4
pre-commit>=3.6
